actions: Split out zerver.actions.user_groups.
(cherry picked from commit 372c10f5f3505711bf592bb03172abbc11f64544)
